BuffKarte und BuffkarteImpl in Kartentypen verändert dadurch BuffCards.java und BuffCardsImpl.java gelöscht
