--- app/src/main/java/com/example/munchkin/Karte/BuffCards.java
@@ -1,6 +1,0 @@
-package com.example.munchkin.Karte;
-
-public interface BuffCards extends Karte {
-    void monsterLvlUp(int montsterLevel);
-    void moreDamageAfterBattle(int battleDamage);
-}
